Move UiPredicates class to client-config.js
Ref b/228856600.
