+++ src/model/client-config.js
@@ -21,7 +21,7 @@
  *   attributionParams: (./attribution-params.AttributionParams|undefined),
  *   autoPromptConfig: (./auto-prompt-config.AutoPromptConfig|undefined),
  *   paySwgVersion: (string|undefined),
- *   uiPredicates: (./auto-prompt-config.UiPredicates|undefined),
+ *   uiPredicates: (UiPredicates|undefined),
  *   usePrefixedHostPath: (boolean|undefined),
  *   useUpdatedOfferFlows: (boolean|undefined),
  *   skipAccountCreationScreen: (boolean|undefined),
@@ -60,10 +60,35 @@ export class ClientConfig {
     /** @const {boolean} */
     this.skipAccountCreationScreen = skipAccountCreationScreen || false;
 
-    /** @const {./auto-prompt-config.UiPredicates|undefined} */
+    /** @const {UiPredicates|undefined} */
     this.uiPredicates = uiPredicates;
 
     /** @const {./attribution-params.AttributionParams|undefined} */
     this.attributionParams = attributionParams;
+  }
+}
+
+/**
+ * Predicates to control UI elements.
+ */
+export class UiPredicates {
+  /**
+   * @param {boolean|undefined} canDisplayAutoPrompt
+   * @param {boolean|undefined} canDisplayButton
+   * @param {boolean|undefined} purchaseUnavailableRegion
+   */
+  constructor(
+    canDisplayAutoPrompt,
+    canDisplayButton,
+    purchaseUnavailableRegion
+  ) {
+    /** @const {boolean|undefined} */
+    this.canDisplayAutoPrompt = canDisplayAutoPrompt;
+
+    /** @const {boolean|undefined} */
+    this.canDisplayButton = canDisplayButton;
+
+    /** @const {boolean|undefined} */
+    this.purchaseUnavailableRegion = purchaseUnavailableRegion;
   }
 }
